Test Setup
This commit adds the necessary cmake changes to execute tests. It also adds two tests, one simple example test, as well as a test which can be run manually on the pi. 
Further a clang-format file is added, which can be used with the pre-commit script to check whether your source files are correctly formatted.

Signed-off-by: Jakob Arndt jakob.arndt@campus.tu-berlin.de